refactor waitgroupx in terms of errgroupx, rename groupx
